Corrected mistakes with the vertex stuff
Viewport runs from 0,0 to viewport.width,viewport.height

I had it running from -viewport.width/2,-viewport.height/2 to
viewport.width/2,viewport.height/2 before, so that the sun was at 0,0.
Not sure if that would be any better.

Made the sun bigger! :)

The vertices are based around 0,0 in their coordinate system and are
translated to the correct position in the world by returning a Matrix,
which is used by the Effect.

Now that all is merged in and I don't have random branches, I'll try to
keep the branches right.