Actually implemented the GravitySimulator
I had broken master branch (oops) in my previous commit as I didn't realise I was committing unfinished work.
I fixed that, and the calculation of gravity is now nicely abstracted away in a separate class.
Currently it is not doing anything different to what the ApplyGravity method used to do, but that could change.
It's also a patten of how I think we could do the collision detection, and other physics things.