Moved _gameObjects out of SpaceWar2Game into GameObjectFactory
This just seemed like a more useful place for it to be. When you create objects using the GameObjectFactory, they are added to its internal _gameObjects collection. This is exposed through a public property (although in the future we might not need it at all...)

I also got rid of the individual references to _sun/_ship1/_ship2 in SpaceWar2Game as they are not needed anymore. GravitySimulator has its own references to any gravity objects, and and other stuff is all done through the GameObjects collection.