Removed most references to viewport - not really required.
The model of the game area (solar system) shold be independent of how it is drawn. We already have a separate camera, that is able to move around the world freely. Its view is then projected onto the screen. So it doesn't make any sense to tie anything into the screen's dimensions. I've left ScreenConstraint() working in practically the same way it did before, but I expect this will change too.