Fixed strange forces behaviour.
Now when a force is applied to the GameObject it is placed in a private queue (well a list) called queuedForces. During the Update loop, a private method ResolveForces() is called. This method adds the queued forces together, and moves them into another (public) list Forces. Both the Forces and ResolvedForce can be queried as many times as required, and are updated each game loop.