Added model support.
Created a new DrawingManager class.
This, like the other managers, maintains a list of IGameObjects.
The GameObjectFactory has a reference to a DrawingManager and registers the Suns and Ships.
It also now has a reference to the ContentManager, so it can load/assign the Models to the new objects.
Have edited the properties of the Sun and Saucer .fbx files, so that they're near the size of the hand drawn versions.
They're not exact, and the 'radius' of an object (used for the area of no gravity) is still hard coded for now.
The color being passed for the lineColor is being used as a Color property, which is in turn being
used by the model drawing method as a parameter to the BasicEffect as diffuse color.
BasicEffect is being used at present, rather than the custom Effect.
DrawingManager has a settable ActiveCamera. It might be an idea to have
a number of Cameras, selectable by keyboard (say [ and ]).

TODO: if the radius can be removed and replaced with something using the Model's BoundingSphere,
that would be good.
TODO: the arrows are not Models yet, there is a .fbx sat there though.

Feel free to pick this commit apart! The code is a mess, some of which I know
and will make better, but some of it is probably just poor coding style :)