Added an EnergyStore ship component with associated tests.
This backs the RequestEnergy function and moves some code
outside the Ship class.

I liked the way that Shield moving out made the testing work
more easily, so I did the same with the Energy side too :)

I think the tests are pretty full. They could do with a check
though please Mr W.

There is some duplication in the code for a Shield and an EnergyStore
in terms of the storage part (i.e., the recharge, max level, 0 level handling)
I wonder if this could be lumped together as one Storage base
class and then the specific functionality of damage and energy requests
built on top?

Oh, and I wrote the tests before 99% of the class itself. Good eh?