Pass wrapper for GraphicsDevice to Draw methods
Created a wrapper for the GraphicsDevice (used to draw some primitive
stuff) to the draw method. So we ask the object to 'draw yourself, and
here is the graphics device to use'.
This means we don't need to inject GraphicsDeviceManager into
GameObjects via constructors, and simplifies testing the draw methods
(cause we just pass in a mocked out IGraphicsDevice)